feat: 간편 회원가입 페이지 경로 수정 및 UI 적용 완료
